pyproject.toml
[tool.poetry]
name = "tendermint-stream"
version = "0.1.0"
description = ""
authors = ["yihuang <yi.codeplayer@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
fastapi = "^0.62.0"
aiohttp = {extras = ["speedups"], version = "^3.7.3"}
uvicorn = {extras = ["standard"], version = "^0.12.3"}
python-decouple = "^3.3"
fire = "^0.3.1"
ujson = "^4.0.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
